requirements.txt
# Exam Explainer Bot - Dependencies
streamlit>=1.29.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
